fixed uint vs int comparison warning

+++ src/solver_program/solver_module.cc
@@ -91,7 +91,7 @@ int main(int argc, char** argv) {
   }
 
   // check that there no conflict between dimensions of container, and the number of integers within <state>*/
-  if ((partial_board[0]*partial_board[1]) != (state_vector.size() - 2)){
+  if ((partial_board[0]*partial_board[1]) != static_cast<int>(state_vector.size() - 2)){
     if (debug) {
       cout << "Error: state input's dimensions don't match container area" << endl;
     }
